refactor: stage 패키지 이동
