+++ src/main/java/org/codequistify/master/application/player/dto/PlayerStageProgressResponse.java
@@ -1,6 +1,6 @@
 package org.codequistify.master.application.player.dto;
 
-import org.codequistify.master.core.domain.stage.dto.StageCodeDTO;
+import org.codequistify.master.application.stage.dto.StageCodeDTO;
 
 import java.util.List;
