+++ src/main/java/org/codequistify/master/application/stage/dto/GradingResponse.java
@@ -1,0 +1,9 @@
+package org.codequistify.master.application.stage.dto;
+
+public record GradingResponse(
+        Boolean isCorrect,
+        Integer nextIndex,
+        Boolean isLast,
+        Boolean isComposable
+) {
+}
