+++ src/main/java/org/codequistify/master/application/stage/dto/HeatMapDataPoint.java
@@ -1,0 +1,9 @@
+package org.codequistify.master.application.stage.dto;
+
+import java.util.Date;
+
+public record HeatMapDataPoint(
+        Date date,
+        Long count
+) {
+}
