+++ src/main/java/org/codequistify/master/application/stage/dto/PageParameters.java
@@ -1,0 +1,22 @@
+package org.codequistify.master.application.stage.dto;
+
+public record PageParameters(
+        int totalPages,
+        int pageSize,
+        int currentPageIndex,
+        int currentPageOfElement,
+        int totalElement
+) {
+    public static PageParameters of(int totalPages,
+                                    int pageSize,
+                                    int currentPageIndex,
+                                    int currentPageOfElement,
+                                    int totalElement) {
+        return new PageParameters(
+                totalPages,
+                pageSize,
+                currentPageIndex,
+                currentPageOfElement,
+                totalElement);
+    }
+}
