+++ src/main/java/org/codequistify/master/application/stage/dto/QuestionRequest.java
@@ -1,0 +1,14 @@
+package org.codequistify.master.application.stage.dto;
+
+import org.codequistify.master.core.domain.stage.domain.AnswerType;
+
+import java.util.List;
+
+public record QuestionRequest(
+        String title,
+        String description,
+        AnswerType answerType,
+        String correctAnswer,
+        List<String> options
+) {
+}
