+++ src/main/java/org/codequistify/master/application/stage/dto/QuestionResponse.java
@@ -1,0 +1,15 @@
+package org.codequistify.master.application.stage.dto;
+
+import org.codequistify.master.core.domain.stage.domain.AnswerType;
+
+import java.util.List;
+
+public record QuestionResponse(
+        Long questionId,
+        Integer index,
+        String title,
+        String description,
+        AnswerType answerType,
+        List<String> options
+) {
+}
