+++ src/main/java/org/codequistify/master/application/stage/dto/StageActionRequest.java
@@ -1,0 +1,7 @@
+package org.codequistify.master.application.stage.dto;
+
+public record StageActionRequest(
+        String stageCode,
+        Integer questionIndex
+) {
+}
