+++ src/main/java/org/codequistify/master/application/stage/dto/StageCodeDTO.java
@@ -1,0 +1,12 @@
+package org.codequistify.master.application.stage.dto;
+
+import org.codequistify.master.core.domain.stage.domain.StageImageType;
+
+public record StageCodeDTO(
+        StageImageType stageCode,
+        String accessUri
+) {
+    public static StageCodeDTO of(StageImageType stageImageType) {
+        return new StageCodeDTO(stageImageType, "");
+    }
+}
