+++ src/main/java/org/codequistify/master/application/stage/dto/StageCompletionRequest.java
@@ -1,0 +1,8 @@
+package org.codequistify.master.application.stage.dto;
+
+import org.codequistify.master.core.domain.stage.domain.CompletedStatus;
+
+public record StageCompletionRequest(
+        CompletedStatus status
+) {
+}
