+++ src/main/java/org/codequistify/master/application/stage/dto/StageCompletionResponse.java
@@ -1,0 +1,9 @@
+package org.codequistify.master.application.stage.dto;
+
+public record StageCompletionResponse(
+        Integer previousExp,
+        Integer previousLevel,
+        Integer currentExp,
+        Integer currentLevel
+) {
+}
