+++ src/main/java/org/codequistify/master/application/stage/dto/StagePageResponse.java
@@ -1,0 +1,12 @@
+package org.codequistify.master.application.stage.dto;
+
+import java.util.List;
+
+public record StagePageResponse(
+        List<StageResponse> stages,
+        PageParameters pageParameters
+) {
+    public static StagePageResponse of(List<StageResponse> stages, PageParameters pageParameters) {
+        return new StagePageResponse(stages, pageParameters);
+    }
+}
