+++ src/main/java/org/codequistify/master/application/stage/dto/StageRegistryRequest.java
@@ -1,0 +1,15 @@
+package org.codequistify.master.application.stage.dto;
+
+import org.codequistify.master.core.domain.stage.domain.DifficultyLevelType;
+import org.codequistify.master.core.domain.stage.domain.StageGroupType;
+
+import java.util.List;
+
+public record StageRegistryRequest(
+        String title,
+        String description,
+        StageGroupType groupType,
+        DifficultyLevelType difficultyLevel,
+        List<QuestionRequest> questions
+) {
+}
