--- src/main/java/org/codequistify/master/core/domain/stage/dto/GradingResponse.java
@@ -1,9 +1,0 @@
-package org.codequistify.master.core.domain.stage.dto;
-
-public record GradingResponse(
-        Boolean isCorrect,
-        Integer nextIndex,
-        Boolean isLast,
-        Boolean isComposable
-) {
-}
