--- src/main/java/org/codequistify/master/core/domain/stage/dto/HeatMapDataPoint.java
@@ -1,9 +1,0 @@
-package org.codequistify.master.core.domain.stage.dto;
-
-import java.util.Date;
-
-public record HeatMapDataPoint(
-        Date date,
-        Long count
-) {
-}
