--- src/main/java/org/codequistify/master/core/domain/stage/dto/StageActionRequest.java
@@ -1,7 +1,0 @@
-package org.codequistify.master.core.domain.stage.dto;
-
-public record StageActionRequest(
-        String stageCode,
-        Integer questionIndex
-) {
-}
