--- src/main/java/org/codequistify/master/core/domain/stage/dto/StageCompletionRequest.java
@@ -1,8 +1,0 @@
-package org.codequistify.master.core.domain.stage.dto;
-
-import org.codequistify.master.core.domain.stage.domain.CompletedStatus;
-
-public record StageCompletionRequest(
-        CompletedStatus status
-) {
-}
