--- src/main/java/org/codequistify/master/core/domain/stage/dto/StageCompletionResponse.java
@@ -1,9 +1,0 @@
-package org.codequistify.master.core.domain.stage.dto;
-
-public record StageCompletionResponse(
-        Integer previousExp,
-        Integer previousLevel,
-        Integer currentExp,
-        Integer currentLevel
-) {
-}
